--- src/styles/drawing.module.scss
@@ -1,25 +1,0 @@
-@import "variables";
-.lists {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: flex-start;
-  .item {
-    position: relative;
-    background-color: #333;
-    margin: 0.5rem;
-    .image {
-      height: $container-width / 3.5;
-      width: $container-width / 3.5;
-      max-height: 25vw;
-      max-width: 25vw;
-    }
-    .link {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-    }
-  }
-}
