+++ src/styles/portfolio.module.scss
@@ -1,21 +1,72 @@
-.item {
+@import "variables";
+
+.nav {
+  margin: 0 auto 3rem;
+}
+
+.list {
   display: flex;
-  flex-direction: column;
-  justify-content: center;
+  flex-wrap: wrap;
+  justify-content: space-between;
   align-items: stretch;
-  position: relative;
-  background-color: #ffffff;
-  padding: 1rem;
-  margin: 0 0 2rem;
-  border-radius: 0.5rem;
-  .title {
-    margin: 0 0 1rem;
+  .item {
+    position: relative;
+    width: $container-width / 2.2;
+    max-width: 42vw;
+    margin: 0.5rem 0.1rem;
+    max-height: 25vw;
+    background-color: #ffffff;
+    border-radius: 1rem;
+    overflow: hidden;
+    .imagewrapper {
+      display: flex;
+      max-height: 10vw;
+      overflow: hidden;
+      .image {
+        object-fit: cover;
+        width: 100%;
+      }
+    }
+    .contents {
+      width: 100%;
+      padding: 1rem;
+      .title {
+        display: inline-block;
+      }
+    }
+    .link {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+    }
   }
-  .link {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
+}
+
+.drawing {
+  .lists {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: flex-start;
+    .item {
+      position: relative;
+      background-color: #333;
+      margin: 0.5rem;
+      .image {
+        height: $container-width / 3.5;
+        width: $container-width / 3.5;
+        max-height: 25vw;
+        max-width: 25vw;
+      }
+      .link {
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+      }
+    }
   }
 }
