+++ src/styles/portfolionav.module.scss
@@ -1,0 +1,15 @@
+.list {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+  .link {
+    padding: 0.5rem;
+    margin: 0 1rem;
+    background-color: #a8a8a8;
+    border-radius: 0.2rem;
+    &_inactive {
+      background-color: transparent;
+    }
+  }
+}
