--- src/styles/service.module.scss
@@ -1,38 +1,0 @@
-@import "variables";
-.list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: stretch;
-  .item {
-    position: relative;
-    width: $container-width / 2.2;
-    max-width: 42vw;
-    margin: 0.5rem 0.1rem;
-    max-height: 25vw;
-    background-color: #ffffff;
-    border-radius: 1rem;
-    overflow: hidden;
-    .imagewrapper {
-      max-height: 10vw;
-      overflow: hidden;
-      .image {
-        object-fit: cover;
-      }
-    }
-    .contents {
-      width: 100%;
-      padding: 1rem;
-      .title {
-        display: inline-block;
-      }
-    }
-    .link {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-    }
-  }
-}
